Fixed a broken integration test.

+++ NumberSearch.DataAccess/InvoiceNinja/Invoice.cs
@@ -79,7 +79,7 @@ namespace NumberSearch.DataAccess.InvoiceNinja
         public int updated_at { get; set; }
         public object archived_at { get; set; }
         public string invoice_number { get; set; }
-        public int discount { get; set; }
+        public decimal discount { get; set; }
         public string po_number { get; set; }
         public string invoice_date { get; set; }
         public string due_date { get; set; }
